pyproject.toml
[project]
name = "stubb"
version = "0.1"
description = "Structured LLM outputs and evals"
readme = "README.md"
requires-python = ">=3.8"
authors = [{name = "Mark Neumann"}]
license = {text = "Apache-2.0"}
classifiers = [
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "pydantic >= 2.0.2",
    "typer >= 0.9.0",
    "llama-cpp-python >= 0.2.32"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/deneutoy/stubb"
Changelog = "https://github.com/deneutoy/stubb/releases"
Issues = "https://github.com/deneutoy/stubb/issues"
CI = "https://github.com/deneutoy/stubb/actions"


[project.optional-dependencies]
test = [
    "pytest",
    "mkdocs >= 1.4.3",
    "mkdocs-material >= 9.1.18",
    "mkdocstrings >= 0.22.0",
    "mkdocstrings-python >= 1.1.2",
    "ruff >= 0.1.0",
]
